Compute inherited colors for product filters editor and frontend
