+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-filters/set-styles.ts
@@ -1,21 +1,4 @@
-function getClosestColor(
-	element: Element | null,
-	colorType: 'color' | 'backgroundColor'
-): string | null {
-	if ( ! element ) {
-		return null;
-	}
-	const color = window.getComputedStyle( element )[ colorType ];
-	if ( color !== 'rgba(0, 0, 0, 0)' && color !== 'transparent' ) {
-		const matches = color.match( /\d+/g );
-		if ( ! matches || matches.length < 3 ) {
-			return null;
-		}
-		const [ r, g, b ] = matches.slice( 0, 3 );
-		return `rgb(${ r }, ${ g }, ${ b })`;
-	}
-	return getClosestColor( element.parentElement, colorType );
-}
+import { getClosestColor } from './utils/get-closest-color';
 
 export default function setStyles(): void {
 	const container = document.querySelector( '.wc-block-product-filters' );
